ini file reader support buffer

+++ tracker/fdfs_trackerd.c
@@ -231,7 +231,7 @@ int main(int argc, char *argv[])
 
 	g_tracker_thread_count = g_max_connections;
 	if ((result=create_work_threads(&g_tracker_thread_count, \
-		tracker_thread_entrance, (void *)sock, tids, \
+		tracker_thread_entrance, (void *)((long)sock), tids, \
 		g_thread_stack_size)) != 0)
 	{
 		free(tids);
